HAL : Change datatype of mix_ctl cfg to fix adb bootup failure
- Mismatch in datatype in userspace and Kernelspace is leading to device
  bootup failure as invalid input channels and output channels are being set
- Add correct data type for mixer ctl configs to fix this issue

+++ hal/audio_extn/audio_extn.c
@@ -510,7 +510,7 @@ static void set_custom_mtmx_params_v2(struct audio_device *adev,
     char *capture_mixer_name_prefix = "AudStr Capture";
     char *mixer_name_suffix = "ChMixer Cfg";
     char mixer_ctl_name[128] = {0};
-    int chmixer_cfg[5] = {0}, len = 0;
+    long int chmixer_cfg[5] = {0}, len = 0;
     int be_id = -1, err = 0;
 
     be_id = platform_get_snd_device_backend_index(pinfo->snd_device);
